Переместит функции отрисовки. Обновит импорты
