+++ src/utils/utils.js
@@ -25,27 +25,4 @@ const isEnterPress = (evt) => {
   return evt.keyCode === KeyCode.ENTER_KEY_CODE;
 };
 
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export {getRandomArrayItem, getRandomIntegerNumber, getRandomInteger, isEscPress, isEnterPress, RenderPosition, createElement, render};
+export {getRandomArrayItem, getRandomIntegerNumber, getRandomInteger, isEscPress, isEnterPress};
